refactor: mv chain-hub to exos
